+++ Shikkhanobish/Shikkhanobish/Model/Student.cs
@@ -10,7 +10,7 @@
         public int Age { get; set; }
         public string Class { get; set; }
         public string InstitutionName { get; set; }
-        public int RechargedAmount { get; set; }
+        public float RechargedAmount { get; set; }
         public int IsPending { get; set; }
 
         public int TotalTuitionTIme { get; set; }
